phase-d(bun_runtime): why_command — fix Context sig, semver::query::parse, gate PackageManager-stub-blocked bodies

+++ src/runtime/api/bun/h2_frame_parser.rs
@@ -1549,7 +1549,7 @@ impl Stream {
                         client.dispatch_with_extra(
                             JSH2FrameParser::Gc::onStreamEnd,
                             identifier,
-                            JSValue::js_number(self.state as u8),
+                            JSValue::js_number(self.state as u8 as f64),
                         );
                     }
                 }
